requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file ateliersoude.egg-info/requirements.txt ateliersoude.egg-info/requirements.in
#
atomicwrites==1.3.0       # via pytest
attrs==19.1.0             # via pytest
cairocffi==1.0.2          # via cairosvg, weasyprint
cairosvg==2.3.0           # via weasyprint
cffi==1.12.2              # via cairocffi, weasyprint
coverage==4.5.3           # via pytest-cov
cssselect2==0.2.1         # via cairosvg, weasyprint
defusedxml==0.5.0         # via cairosvg
diff-cover==2.0.0
django-bootstrap-static==4.0.0
django-crispy-forms==1.7.2  # via django-fm
django-debug-toolbar==1.11
django-fm==0.2.5
django-simple-history==2.7.0
django==2.2
djangorestframework==3.9.2
entrypoints==0.3          # via flake8
factory-boy==2.11.1
faker==1.0.5              # via factory-boy
flake8==3.7.7             # via pytest-flake8
html5lib==1.0.1           # via weasyprint
inflect==2.1.0            # via jinja2-pluralize
inflection==0.3.1         # via pytest-factoryboy
jinja2-pluralize==0.3.0   # via diff-cover
jinja2==2.10.1            # via diff-cover, jinja2-pluralize
lxml==4.3.3
markupsafe==1.1.1         # via jinja2
mccabe==0.6.1             # via flake8
more-itertools==7.0.0     # via pytest
pillow==6.0.0             # via cairosvg
pluggy==0.9.0             # via pytest
psycopg2==2.8.2
py==1.8.0                 # via pytest
pycodestyle==2.5.0        # via flake8
pycparser==2.19           # via cffi
pyflakes==2.1.1           # via flake8
pygments==2.3.1           # via diff-cover
pyphen==0.9.5             # via weasyprint
pytest-cov==2.6.1
pytest-django==3.4.8
pytest-factoryboy==2.0.2
pytest-flake8==1.0.4
pytest==4.4.0             # via pytest-cov, pytest-django, pytest-factoryboy, pytest-flake8
python-dateutil==2.8.0    # via faker
pytz==2019.1              # via django
six==1.12.0               # via diff-cover, faker, html5lib, pytest, python-dateutil
sqlparse==0.3.0           # via django, django-debug-toolbar
text-unidecode==1.2       # via faker
tinycss2==1.0.2           # via cairosvg, cssselect2, weasyprint
weasyprint==47
webencodings==0.5.1       # via html5lib, tinycss2
